// frontend/src/generated/api/types.gen.ts
export type ArticlesResponse = {
    articles: Array<{
        slug: string;
        title: string;
        description: string;
        tagList: Array<string>;
        createdAt: string;
        updatedAt: string;
        favorited: boolean;
        favoritesCount: number;
        author: {
            username: string;
            bio: string | null;
            image: string | null;
            following: boolean;
        };
    }>;
    articlesCount: number;
};
// ...
export type UserResponse = {
    user: {
        email: string;
        token: string;
        username: string;
        bio: string | null;
        image: string | null;
    };
};